// src/State/types/orderType.tsx
import { Deal } from "./DealTypes";
import { Product } from "./ProductType";
import { Address, User } from "./userTypes";

export interface OrderState {
    orders: Order[];
    orderItem:OrderItem | null;
    currentOrder: Order | null;
    paymentOrder: any | null;
    loading: boolean;
    error: string | null;
    orderCanceled: boolean
}

export interface Order{
  createdAt: any;
  items: any;
     paymentMethod: string;
    id: number;
    orderId: string;
    user: User;
    sellerId: number;
    orderItems: OrderItem[];
    orderDate: string;
    shippingAddress: Address;
    paymentDetails: any;
    totalMrpPrice: number;
    totalSellingPrice?: number;
    discount?: number;
    orderStatus: OrderStatus;
    totalItem: number;
    deliverDate: string;
}

export enum OrderStatus {
    PENDING = 'PENDING',
    SHIPPED = 'SHIPPED',
    DELIVERED = 'DELIVERED',
    CANCELLED = 'CANCELLED'
}
 export interface OrderItem {
    id: number;
    orderId: number; 
    product: Product;
    deal?: Deal;
    size: string;
    quantity: number;
    mrpPrice: number;
    sellingPrice: number;
 }